디테일작업 진행중

+++ Strong1team2/src/main/java/com/smhrd/controller/UnregisterService.java
@@ -14,58 +14,33 @@ public class UnregisterService implements Command {
 
 	public String execute(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		
+
 		HttpSession session = request.getSession();
+		MemberVO mem = (MemberVO) session.getAttribute("result");
 		System.out.println("----------------------------중간점검1----------------------------");
 		String email = request.getParameter("email");
-		
-		
+		String pw = request.getParameter("pw");
+
 		MemberVO vo = new MemberVO();
 		vo.setEmail(email);
-		
-		System.out.println(email);
-		
-		
-		
-		DAO dao = new DAO();
-		System.out.println("dao 생성");
-		int row = dao.delete(vo);
-		
-		System.out.println("----------------------------중간점검2----------------------------");
 
-		if (row > 0) {
-			System.out.println("----------------------------중간점검3----------------------------");
+		System.out.println(email);
+
+		DAO dao = new DAO();
+
+		if (mem.getEmail().equals(email) && mem.getPw().equals(pw)) {
+			int row = dao.delete(vo);
+
+			if (row > 0) {
 			
-			session.invalidate();
-			return "redirect:/Gomain.do";
-		} else {
-			System.out.println("실패");
-			// redirect:/ >> 우리의 약속 기호!!
-			// >> redirect 방식으로 이동해라 !! 를 FC에게 알려주는 역할!!
-			return "redirect:/Gomodify.do";
-			
+				session.invalidate();
+				return "redirect:/Gomain.do";
+			}
+
 		}
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
+
+		return "redirect:/Gounregister.do?error=invalidCredentials";
 
 	}
+
 }
